back-end/requirements.txt
beautifulsoup4==4.11.1
Django==4.1.1
djangorestframework==3.14.0
fake_useragent==1.1.1
geotext==0.3.0
pytz==2022.2.1
requests==2.28.1
requests_html==0.10.0
selenium==4.8.0
